Add Review constructor and re-orgonize code
